Final Changes

+++ public/delete.php
@@ -66,23 +66,23 @@
 <!-- This is a loop, which will loop through each result in the array -->
 <?php foreach($result as $row) { ?>
             
-
-
         <p>
             <br>
             ID: 
-            <?php echo escape($row['id']); ?><br> Unit Name:
-            <?php echo $row['unitname']; ?><br> Due Date:
-            <?php echo $row['duedate']; ?><br> Assignment Name:
-            <?php echo $row['assignmentname']; ?><br> 
-            <br><a href='delete.php?id=<?php echo $row['id']; ?>'><button class= "btn btn-danger" data-toggle="modal" data-target="delete">Delete <span class="glyphicon glyphicon-trash"</span></button></a>
-                
+    <?php echo $row['id']; ?><br> Unit Name:
+    <?php echo $row['unitname']; ?><br> Due Date:
+    <?php echo $row['duedate']; ?><br> Assignment Name:
+    <?php echo $row['assignmentname']; ?><br> Priority:
+    <?php echo $row['priority']; ?><br> 
+            
+            
+    <br><a href='delete.php?id=<?php echo $row['id']; ?>'><button class= "btn btn-danger" onclick="return confirm('Are you sure you want to delete this item')" data-toggle="modal" data-target="delete">Delete <span class="glyphicon glyphicon-trash"</span></button></a>
+                   
         </p>
 
 
 <hr>
-<?php }; //close the foreach
-   
+<?php }; 
 ?>
 
 <!DOCTYPE html>
@@ -97,42 +97,5 @@
   <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/js/bootstrap.min.js"></script>
 </head>
 <body>
-
-   
     
-<!-- Modal -->
-<div class="modal modal-danger fade" id="delete" tabindex="-1" role="dialog" aria-labelledby="myModalLabel">
-  <div class="modal-dialog" role="document">
-    <div class="modal-content">
-      <div class="modal-header">
-        <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
-        <h4 class="modal-title text-center" id="myModalLabel">Delete Confirmation</h4>
-      </div>
-      <form action="{{route('category.destroy','test')}}" method="post">
-      		{{method_field('delete')}}
-      		{{csrf_field()}}
-	      <div class="modal-body">
-				<p class="text-center">
-					Are you sure you want to delete this?
-				</p>
-	      		<input type="hidden" name="category_id" id="cat_id" value="">
-
-	      </div>
-	      <div class="modal-footer">
-	        <button type="button" class="btn btn-success" data-dismiss="modal">No, Cancel</button>
-	        <button type="submit" class="btn btn-warning">Yes, Delete</button>
-	      </div>
-      </form>
-    </div>
-  </div>
-</div>
-
-</body>
-</html>
-  
-
-
-
-
-
 <?php include "templates/footer.php"; ?>
